+++ waku/v2/node/storage/migration/migrations_scripts/message/00003_deleteReceiverTimeStamp.up.sql
@@ -1,0 +1,2 @@
+ALTER TABLE Message IF EXISTS
+DROP COLUMN receiverTimestamp;